\name{gwb-term}
\alias{gwb}
\alias{gwb1nsp}
\alias{gwb2nsp}
\docType{methods}
\title{Bipartite gwb1nsp and gwb2nsp terms}
\description{
  These terms work with edgewise shared partners in undirected bipartite networks. 
}
%\usage{
%terms.ergm()

%}
%\arguments{
%\item{\dots}{Specific to the model term.}
%}

\details{
details needed

\code{gwb1nsp(alpha=0, fixed=FALSE, cutoff=30)}

\code{gwb2nsp(alpha=0, fixed=FALSE, cutoff=30)}

Note that allthough the default is \code{fixed=FALSE} The terms are not yet able to handle a non-fixed decay term so it must be set to TRUE.
}


%\value{\code{\link{ergm}} returns an object of class 'ergm' that is a list.
%}
\references{
References? 

}

\author{
Matthew Hamilton  \email{matthewlhamilton@gmail.com}
}

\examples{

# example needed
}
\keyword{models}
