\name{ergm.terms.contrib}
\alias{ergm.terms.contrib}
\docType{package}
\title{
  A collection of ERGM change-statistics contributed by the statnet user community
}
\description{
  \pkg{ergm.terms.contrib} is collection of \pkg{ergm} change-statistics (model terms) 
  that have been coded and contributed by members \pkg{statnet} user community using 
  the \pkg{ergm.userterms} package as template. 
   
  \pkg{statnet} is a collection of software packages for statistical network 
  analysis. The packages implement recent advances in network modeling based on
  exponential-family random graph models (ERGM).  

}

\details{
  Terms in this package can be used in ergm models by loading the package with \code{library(ergm.terms.contrib)} 
  Please see the documentation for each of the individual terms in the package for specific details and author information.  The help page \code{?ergm.terms} provides more general information.  A tutorial on writing user terms is availible at \url{https://statnet.org/trac/raw-attachment/wiki/Resources/ergmusertermsPaper.pdf}
  
  If you have a term that you would like to contribute to this package, please submit a pull request to \url{https://github.com/statnet/ergm.terms.contrib} or email the package maintainers. 
    
}

\author{
  Maintainer: Martina Morris \email{morris@uw.edu}
  Authors for specific terms are listed on the help page for each term.
}
